fix spring webflux path extraction + improve webflux tests
