+++ dd-java-agent/instrumentation/spring-webflux-5/build.gradle
@@ -1,4 +1,3 @@
-
 muzzle {
   pass {
     name = "webflux_5.0.0+_with_netty_0.8.0"
@@ -44,14 +43,16 @@ testSets {
     dirName = 'test'
   }
 
-  boot20Test {}
   latestBoot20Test {
-    dirName = 'boot20Test'
+    dirName = 'bootTest'
   }
 
-  boot24Test {}
   latestBoot24Test {
-    dirName = 'boot24Test'
+    dirName = 'bootTest'
+  }
+
+  latestBoot2LatestTest {
+    dirName = 'bootTest'
   }
 }
 
@@ -73,28 +74,22 @@ dependencies {
   latestDepTestImplementation group: 'io.projectreactor.netty', name: 'reactor-netty', version: '0.+'
   latestDepTestImplementation group: 'org.springframework', name: 'spring-test', version: '5.+'
 
-  boot20TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-webflux', version: '2.0.0.RELEASE'
-  boot20TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.0.0.RELEASE'
-  boot20TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-reactor-netty', version: '2.0.0.RELEASE'
 
   latestBoot20TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-webflux', version: '2.0.+'
   latestBoot20TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.0.+'
   latestBoot20TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-reactor-netty', version: '2.0.+'
 
-  boot24TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-webflux', version: '2.4.0'
-  boot24TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.4.0'
-  boot24TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-reactor-netty', version: '2.4.0'
-
   latestBoot24TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-webflux', version: '2.4.+'
   latestBoot24TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.4.+'
   latestBoot24TestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-reactor-netty', version: '2.4.+'
+
+  latestBoot2LatestTestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-webflux', version: '2+'
+  latestBoot2LatestTestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2+'
+  latestBoot2LatestTestImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-reactor-netty', version: '2+'
 }
 
-tasks.named("test").configure {
-  dependsOn "boot20Test"
-  dependsOn "boot24Test"
-}
 tasks.named("latestDepTest").configure {
   dependsOn "latestBoot20Test"
   dependsOn "latestBoot24Test"
+  dependsOn "latestBoot2LatestTest"
 }
