--- dd-java-agent/instrumentation/spring-webflux-5/src/boot24Test/groovy/dd/trace/instrumentation/springwebflux/server/EchoHandler.groovy
@@ -1,21 +1,0 @@
-package dd.trace.instrumentation.springwebflux.server
-
-import datadog.trace.api.Trace
-import org.springframework.http.MediaType
-import org.springframework.stereotype.Component
-import org.springframework.web.reactive.function.server.ServerRequest
-import org.springframework.web.reactive.function.server.ServerResponse
-import reactor.core.publisher.Mono
-
-/**
- * Note: this class has to stay outside of 'datadog.*' package because we need
- * it transformed by {@code @Trace} annotation.
- */
-@Component
-class EchoHandler {
-  @Trace(operationName = "echo", resourceName = "echo")
-  Mono<ServerResponse> echo(ServerRequest request) {
-    return ServerResponse.accepted().contentType(MediaType.TEXT_PLAIN)
-      .body(request.bodyToMono(String), String)
-  }
-}
