--- dd-java-agent/instrumentation/spring-webflux-5/src/boot24Test/groovy/dd/trace/instrumentation/springwebflux/server/FooModel.groovy
@@ -1,16 +1,0 @@
-package dd.trace.instrumentation.springwebflux.server
-
-class FooModel {
-  public long id
-  public String name
-
-  FooModel(long id, String name) {
-    this.id = id
-    this.name = name
-  }
-
-  @Override
-  String toString() {
-    return "{\"id\":" + id + ",\"name\":\"" + name + "\"}"
-  }
-}
