+++ dd-java-agent/instrumentation/spring-webflux-5/src/bootTest/groovy/SpringWebfluxTest.groovy
@@ -65,8 +65,7 @@ class SpringWebfluxTest extends AgentTestRunner {
             "$Tags.HTTP_STATUS" 200
             "$Tags.HTTP_USER_AGENT" String
             "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-            // BUG
-            if (!testName.startsWith("functional")) { "$Tags.HTTP_ROUTE" "$urlPathWithVariables" }
+            "$Tags.HTTP_ROUTE" "$urlPathWithVariables"
             defaultTags()
           }
         }
@@ -143,8 +142,7 @@ class SpringWebfluxTest extends AgentTestRunner {
             "$Tags.HTTP_STATUS" 200
             "$Tags.HTTP_USER_AGENT" String
             "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-            // BUG
-            if (!testName.startsWith("functional")) { "$Tags.HTTP_ROUTE" "$urlPathWithVariables" }
+            "$Tags.HTTP_ROUTE" "$urlPathWithVariables"
             defaultTags()
           }
         }
@@ -349,7 +347,7 @@ class SpringWebfluxTest extends AgentTestRunner {
             "$Tags.HTTP_STATUS" 202
             "$Tags.HTTP_USER_AGENT" String
             "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-            // BUG "$Tags.HTTP_ROUTE" "/echo"
+            "$Tags.HTTP_ROUTE" "/echo"
             defaultTags()
           }
         }
@@ -411,8 +409,7 @@ class SpringWebfluxTest extends AgentTestRunner {
             "$Tags.HTTP_STATUS" 500
             "$Tags.HTTP_USER_AGENT" String
             "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-            // BUG
-            if (!testName.startsWith("functional")) { "$Tags.HTTP_ROUTE" "$urlPathWithVariables" }
+            "$Tags.HTTP_ROUTE" "$urlPathWithVariables"
             defaultTags()
           }
         }
@@ -489,7 +486,7 @@ class SpringWebfluxTest extends AgentTestRunner {
             "$Tags.HTTP_STATUS" 307
             "$Tags.HTTP_USER_AGENT" String
             "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-            // BUG "$Tags.HTTP_ROUTE" "/double-greet-redirect"
+            "$Tags.HTTP_ROUTE" "/double-greet-redirect"
             defaultTags()
           }
         }
@@ -527,7 +524,7 @@ class SpringWebfluxTest extends AgentTestRunner {
             "$Tags.HTTP_STATUS" 200
             "$Tags.HTTP_USER_AGENT" String
             "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-            // BUG "$Tags.HTTP_ROUTE" "/double-greet"
+            "$Tags.HTTP_ROUTE" "/double-greet"
             defaultTags()
           }
         }
@@ -581,8 +578,7 @@ class SpringWebfluxTest extends AgentTestRunner {
               "$Tags.HTTP_STATUS" 200
               "$Tags.HTTP_USER_AGENT" String
               "$Tags.HTTP_CLIENT_IP" "127.0.0.1"
-              // BUG
-              if (!testName.startsWith("functional")) { "$Tags.HTTP_ROUTE" "$urlPathWithVariables" }
+              "$Tags.HTTP_ROUTE" "$urlPathWithVariables"
               defaultTags()
             }
           }
